Move identity provider into sso package
