+++ backend/app/adapter/facebook/identityprovider.go
@@ -5,7 +5,7 @@ import (
 	"net/url"
 
 	"github.com/short-d/app/fw/webreq"
-	"github.com/short-d/short/backend/app/usecase/external"
+	"github.com/short-d/short/backend/app/usecase/sso"
 )
 
 // More info here: https://developers.facebook.com/docs/facebook-login/manually-build-a-login-flow
@@ -17,7 +17,7 @@ const (
 	fbResponseType     = "code"
 )
 
-var _ external.IdentityProvider = (*IdentityProvider)(nil)
+var _ sso.IdentityProvider = (*IdentityProvider)(nil)
 
 // IdentityProvider represents Facebook OAuth service.
 type IdentityProvider struct {
